refactor : 메서드 분리 및 패키지 이동
