+++ src/main/kotlin/blackjack/domain/player/Dealer.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.player
 
-import blackjack.GameStatus
+import blackjack.domain.blackjackgame.GameStatus
 import blackjack.domain.card.Card
 import blackjack.state.Stand
