+++ src/main/kotlin/blackjack/domain/player/UserRole.kt
@@ -1,6 +1,6 @@
 package blackjack.domain.player
 
-import blackjack.GameStatus
+import blackjack.domain.blackjackgame.GameStatus
 import blackjack.domain.card.Card
 
 abstract class UserRole(
